phase-2-web/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
dependencies = [
    "fastapi[standard]",
    "sqlmodel",
    "pydantic-settings",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "alembic",
    "psycopg2-binary",
]

